app/requirements.txt
fastapi
uvicorn
pydantic
flask
python-nmap
httpx
